# apps/backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/send-otp": {
      "post": {
        "operationId": "AuthController_sendOtp_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP accepted for delivery",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "message": "OTP sent successfully"
                    },
                    "timestamp": "2026-03-21T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "429": {
            "description": "OTP rate limit exceeded"
          }
        },
        "summary": "Send OTP to user phone",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyOtp_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified and access_token cookie issued",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "user": {
                        "id": "user_123",
                        "phone": "+919876543210",
                        "role": "BUYER",
                        "fullName": "Buyer User"
                      },
                      "onboarding": {
                        "completed": true,
                        "nextStep": null
                      }
                    },
                    "timestamp": "2026-03-21T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "429": {
            "description": "OTP verification rate limit exceeded"
          }
        },
        "summary": "Verify OTP and establish session cookie",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Session refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "message": "Session refreshed"
                    },
                    "timestamp": "2026-03-21T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access_token cookie"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Refresh access token cookie using authenticated session",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Authenticated user context",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "sub": "user_123",
                    "phone": "+919876543210",
                    "role": "VENDOR",
                    "hasVendorProfile": true
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access_token cookie"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get authenticated user payload from JWT",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "BuildMart API",
    "description": "BuildMart construction procurement platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access_token": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
...
      "SendOtpDto": {
        "type": "object",
        "properties": {}
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}